% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.S.fr.R
\name{CV.S.fr}
\alias{CV.S.fr}
\title{The cross-validation (CV) score}
\usage{
CV.S.fr(
  y,
  S,
  W = NULL,
  trim = 0,
  andf = 0,
  draw = FALSE,
  metric = metric.lp,
  par.metric = list()
)
}
\arguments{
\item{y}{Matrix of set cases with dimension (\code{n} x \code{m}), where
\code{n} is the number of curves and \code{m} are the points observed in
each curve.}

\item{S}{Smoothing matrix, see \code{\link{S.NW}}, \code{\link{S.LLR}} or
\eqn{S.KNN}.}

\item{W}{Matrix of weights.}

\item{trim}{The alpha of the trimming.}

\item{andf}{Degree of freedom}

\item{draw}{=TRUE, draw the curves, the sample median and trimmed mean.}

\item{metric}{Metric function, by default \code{\link{metric.lp}}.}

\item{par.metric}{Further arguments passed to metric.}
}
\value{
{ Returns CV score calculated for input parameters.  }
}
\description{
Compute the leave-one-out cross-validation score.
}
\seealso{
See Also as \code{\link{fregre.np.cv.fr}} \cr Alternative method:
\code{\link{GCV.S.fr}}
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{utilities}
