% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata2basis.R
\name{fdata2basis}
\alias{fdata2basis}
\alias{summary.basis.fdata}
\title{Compute the representation in a basis of a curve}
\usage{
fdata2basis(fdataobj, basis, method = c("grid", "inprod"), center = TRUE)

\method{summary}{basis.fdata}(object, draw = TRUE, index = NULL, ...)
}
\arguments{
\item{fdataobj}{\code{\link{fdata}} class object.}

\item{basis}{a functional basis object defining the basis}

\item{method}{character string, if it is "grid" the fdata object is evaluated in the grid (\code{argvals} of fdata),
if it is "inprod" the basis representation of functional data is computed by inner products 
(\code{\link{inprod.fdata}(fdataobj,basis)}).}

\item{center}{a logical value,  if  \code{TRUE} (by default), the function 
centres the curves by subtracting the functional mean}

\item{object}{\code{basis.fdata} class object calculated by: \code{\link{fdata2basis}}}

\item{draw}{logical, original curves  and their basis representation are plotted}

\item{index}{vector, by default (if NULL) the first n curves are plotted, where n = min(4, length(fdataobj)). 
Otherwise, index vector indicates taht curvesare plotted.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
fdata2basis \code{fdata2bais} function return: 
\itemize{
\item {coefs}{ a matrix with the expansion coefficients for each curve.}
\item {basis}{ basis of \code{\link{fdata}} class evaluated in the same grid of \code{fdataobj}.} 
\item {fdataobj}{ the original \code{fdataobj} object.} 
\item {mean}{ functional mean of \code{fdataobj}.} 
}

\code{summary} function return: 
\itemize{
\item {fdata.est}{ The reconstruction of \code{fdataobj} in the \code{basis} expansion}
\item {R}{ A vector with a measure similar to R-sq for each curve aproximation. }
}
}
\description{
Compute the functional coefficients in a (fixed or data-driven) basis of a curve (\code{\link{fdata}} class object)
}
\examples{
\dontrun{
T <- 71
S <- 51
tj <- round(seq(0,1,len=T),3)
si <- round(seq(0,1,len=S),3)
beta1 <- outer(si,tj,function(si,tj){exp(-5*abs((tj-si)/5))})
nbasis.s =7
nbasis.t=11
base.s <- create.fourier.basis(c(0,1),nbasis=nbasis.s)
base.t <- create.fourier.basis(c(0,1),nbasis=nbasis.t)
y1 <- fdata(rbind(log(1+tj),1-5*(tj-0.5)^2),argvals=tj,rangeval=c(0,1))
aa <- fdata2basis(y1,base.t,method="inprod")
summary(aa)
plot(gridfdata(aa$coefs,aa$basis))
lines(y1,lwd=2,col=c(3,4),lty=2)
}
}
\seealso{
Inverse function: \code{\link{gridfdata}}.
Alternative method: \code{\link{fdata2pc}}, \code{\link{fdata2pls}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente  \email{manuel.oviedo@usc.es}
}
\keyword{multivariate}
