% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.gsam.con.R
\name{fregre.gsam.con}
\alias{fregre.gsam.con}
\title{Fitting Functional Concurrent (Generalized) Spectral Additive Models}
\usage{
fregre.gsam.con(formula, data = list(), h = 1, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced
to that class): a symbolic description of the model to be fitted. The
details of model specification are given under \code{Details}.}

\item{data}{\code{list}, \code{ldata} or \code{mfdata} class object that containing the variables in the model.}

\item{h}{\code{integer}, if \eqn{h=0} (by default) same discretization point for the functional
response and the functional covariates is, for example \eqn{Y(t_j)=f(X(t_j))+\varepsilon}. 
if \eqn{h>0}, the values observed in the ball are also used (j-h,j+h),
 for example \eqn{Y(t_j)=f(X(t_{j-1}),X(t_j),X(t_{j+1})+\varepsilon} with \eqn{h=1}.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Return \code{gam} object plus:
\itemize{
\item {fitted.values}{ \code{fdata} object with the tted model predictions of expected value for each datum.} 
\item {residuals}{ \code{fdata} object with the model residuals.} 
\item {coefficients}{\code{matrix}, with coefficients of each fitted
 model (by row) and parametric coefficients (by col).}
\item {edf.coef}{\code{matrix},  with estimated degrees of freedom 
for each model fitted model (by row) and parametric coefficients (by col).
the NA value is assigned to non-significant linear and smooth terms (p-value> 0.05).} 
\item{result}{\code{list}, with each fitted \code{\link{fregre.gsam}} fitted model
for each dimension of functional response basis representation.}
\item {data}{ List that containing the variables in the model.} 
\item {formula}{ formula.} 
}
}
\description{
Computes functional GAM model between functional covariate
\eqn{(X^1(t_1),\cdots,X^{q}(t_q))}{(X(t_1),...,X(t_q))} (and non functional
covariate \eqn{(Z^1,...,Z^p)}{(Z1,...,Zp)}) and functional response
 \eqn{Y(s)=\left[ Y^{1}(s_1),\cdots,Y^{p}(s_p)\right]}{Y(s)=[Y^1(s),...,Y^r(s)}.

This function is an extension of the functional generalized additve
regression model \code{\link{fregre.gsam}} where now the response is of functional nature.

\deqn{E[Y|X,Z]) = \sum_{l=1}^{L}{f_{l}(\theta_l)}}{E[Y|X,Z]=\sum_l\sum_{l=1}{f_j(\theta_l)}} where
\eqn{\theta_l}{\xi_l}   is the coefficient of the basis function expansion of Y,  
\eqn{\xi_j^k}{\xi_j^k} is the coefficient of the basis function expansion of
\eqn{X^k}, (in PCA analysis \eqn{\xi_j^k}{\xi_j^k} is the score of the
\eqn{j}-functional PC of \eqn{X^k}) 

 \deqn{E[Y|X,Z]) = \alpha+\sum_{i=1}^{p}f_{i}(Z^{i})+\sum_{k=1}^{q}\sum_{j=1}^{k_q}{f_{j}^{k}(\xi_j^k)}}{E[Y|X,Z]=\alpha+\sum_i f_i(Z_{i})+\sum_k^q\sum_{j=1}^{k_q}{f_j^k(\xi_j^k)}}


Note that if \code{basis.y} is null, the coefficients of the basis are 
the raw values of the response observed in the grid (argvas of Y).

The smooth functions \eqn{f(\cdot)}{f(.)} can be added to the right hand
side of the formula to specify that the linear predictor depends on smooth
functions of predictors using smooth terms \code{\link{s}} and
\code{\link{te}} as in \code{\link{gam}} (or linear functionals of these as
\eqn{Z\beta} and \eqn{\big<X(t),\beta\big>}{< X(t),\beta(t) >} in
\code{\link{fregre.glm}}).

The first item in the \code{data} list is called \emph{"df"} and is a data
frame with the non functional explanatory variables, as
\code{\link{gam}}.\cr

Functional covariates and functional response of class \code{fdata} are introduced in
the following items in the \code{data} list.\cr
\code{basis.x} is a list of basis for represent each functional covariate. The basis object can be
created by the function: \code{\link{create.pc.basis}} or  \code{\link{create.fdata.basis}}.

\code{basis.y} is a  basis for represent the functional response.
 \itemize{
\item If it is \code{NULL} (by default) the response variable is not represented in any base. 
\item If it is fixed basis (\code{basisfd} class object) or data-driven basis (\code{fdata.comp}
 class for functional PCA and PLS basis)
}
}
\details{
A \code{\link{fregre.gsam}} model is adjusted for
each of the coefficients functions derived from the basis representation on a functional response.
}
\examples{
\dontrun{
data(aemet)
class(aemet)<-c("ldata","list")
set.seed(0)
ii <- sample(nrow(aemet$df),60)
ltrain <- aemet[ii,row=T]
ltest <- aemet[-ii,row=T]
ff <- logprec~+s(temp,k=3)+s(wind.speed,k=3)#+s(latitude,k=3)
ff <- logprec~+s(temp,k=3)+s(latitude,k=3)
res0=fregre.gsam.con(ff,ltrain,h=0)
res1=fregre.gsam.con(ff,ltrain,h=1)
res2=fregre.gsam.con(ff,ltrain,h=2)
res3=fregre.gsam.con(ff,ltrain,h=3)
summary(res0$result[[151]])
summary(res2$result[[1]])
newy<-ltest$logprec
y <- ltrain$logprec
pred0 <- predict.fregre.gsam.con(res0,ltest)
pred1 <- predict.fregre.gsam.con(res1,ltest)
pred2 <- predict.fregre.gsam.con(res2,ltest)
pred3 <- predict.fregre.gsam.con(res3,ltest)
rest=sum(norm.fdata(newy -func.mean(y))^2)
1-sum(norm.fdata(newy -pred0)^2)/rest
1-sum(norm.fdata(newy -pred1)^2)/rest
1-sum(norm.fdata(newy -pred2)^2)/rest
1-sum(norm.fdata(newy -pred3)^2)/rest
plot(newy -pred0,col=1)
lines(newy-pred1,col=2)
lines(newy-pred2,col=3)
lines(newy-pred3,col=4)

data(tecator)
x=tecator$absorp.fdata
x.d1<-fdata.deriv(x)
x.d2<-fdata.deriv(x.d1)
tt<-x[["argvals"]]
dataf=as.data.frame(tecator$y)
ldat=ldata("df"=dataf,"x"=x,"x.d1"=x.d1,"x.d2"=x.d2)
ltrain <- ldat[1:186,row=T]
ltest <- ldat[-(1:186),row=T]
res0=fregre.gsam.con(x.d2~+s(x),ltrain,h=0)
res1=fregre.gsam.con(x.d2~+s(x),ltrain,h=1)
res2=fregre.gsam.con(x.d2~+s(x),ltrain,h=2)
res3=fregre.gsam.con(x.d2~+s(x),ltrain,h=3)
res4=fregre.gsam.con(x.d2~+s(Water)+s(x),ltrain,h=0)
res5=fregre.gsam.con(x.d2~+s(Water)+s(x),ltrain,h=1)
res6=fregre.gsam.con(x.d2~+s(Water)+s(x),ltrain,h=2)
res7=fregre.gsam.con(x.d2~+s(Water)+s(x),ltrain,h=3) 

newy<-ltest$x.d2
y <- ltrain$x.d2
pred0 <- predict.fregre.gsam.con(res0,ltest)
pred1 <- predict.fregre.gsam.con(res1,ltest)
pred2 <- predict.fregre.gsam.con(res2,ltest)
pred3 <- predict.fregre.gsam.con(res3,ltest)
pred4 <- predict.fregre.gsam.con(res4,ltest)
pred5 <- predict.fregre.gsam.con(res5,ltest)
pred6 <- predict.fregre.gsam.con(res6,ltest)
pred7 <- predict.fregre.gsam.con(res7,ltest)
rest=sum(norm.fdata(newy -func.mean(y))^2)
1-sum(norm.fdata(newy -pred0)^2)/rest; 1-sum(norm.fdata(newy -pred1)^2)/rest
1-sum(norm.fdata(newy -pred2)^2)/rest; 1-sum(norm.fdata(newy -pred3)^2)/rest
1-sum(norm.fdata(newy -pred4)^2)/rest; 1-sum(norm.fdata(newy -pred5)^2)/rest
1-sum(norm.fdata(newy -pred6)^2)/rest; 1-sum(norm.fdata(newy -pred7)^2)/rest 
plot(newy -pred0,col=1)
lines(newy-pred1,col=2)
lines(newy-pred2,col=3)
}

}
\references{
Muller HG and Stadtmuller U. (2005). \emph{Generalized
functional linear models.} Ann. Statist.33 774-805.

Wood (2001) \emph{mgcv:GAMs and Generalized Ridge Regression for R}. R News
1(2):20-25.

Ramsay, James O., and Silverman, Bernard W. (2006), \emph{ Functional Data
Analysis}, 2nd ed., Springer, New York.
}
\seealso{
See Also as: \code{\link{predict.fregre.sam.fr}}.
 Alternative methods: \code{\link{fregre.lm.fr}}
and \code{\link{fregre.kam.fr}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{regression}
