% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.lm.fr.R, R/fregre.mlm.fr.R
\name{fregre.lm.fr}
\alias{fregre.lm.fr}
\alias{fregre.mlm.fr}
\title{Fitting Functional Linear Models with functional response using basis representation.}
\usage{
fregre.lm.fr(
  formula,
  data,
  basis.y = NULL,
  basis.x = NULL,
  lambda = NULL,
  P = NULL,
  ...
)

fregre.mlm.fr(
  formula,
  data,
  basis.y = NULL,
  basis.x = NULL,
  lambda = NULL,
  P = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced
to that class): a symbolic description of the model to be fitted. The
details of model specification are given under \code{Details}.}

\item{data}{List that containing the variables in the model. Allowed objects:
\itemize{
\item A list with \code{\link{fdata}} objects and the "df" object for non-functional covariates.
\item An object of class \code{\link{ldata}}. (equivalent to the previous one)
\item An object of class \code{\link{mfdata}} (only for \code{\link{fdata}} objects).
}}

\item{basis.y}{Basis for functional response variable.}

\item{basis.x}{List of basis for functional explanatory data estimation.}

\item{lambda}{List, indexed by the names of the functional covariates, which contains the Roughness penalty parameter.}

\item{P}{List, indexed by the names of the functional covariates, which contains the parameters for the creation of the penalty matrix.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Return \code{lm} object (MLM output) plus:
\itemize{
\item \code{sr2}{ Residual variance.}
\item \code{basis.y}{ Basis used for response.} 
\item \code{basis.x}{ Basis used for \code{fdata} or \code{fd} covariates.} 
\item \code{data}{ List that containing the variables in the model.}
\item \code{formula}{ formula.}
}
}
\description{
Computes functional regression between functional (and non functional)
explanatory variables and functional response using basis representation.
}
\details{
This section is presented as an extension of the functional linear regression models with scalar response:
\code{\link{fregre.lm}}, \code{\link{fregre.pc}}, \code{\link{fregre.pls}} and
\code{\link{fregre.basis}}. Now, the functional response \eqn{Y} is estimated by
more than one functional covariate \eqn{X^j(t)} and also more than one non
functional covariate \eqn{Z^j}. The regression model is given by:
\deqn{E[Y|X,Z]=\alpha+\sum_{j=1}^{p}\beta_{j}Z^{j}+\sum_{k=1}^{q}\frac{1}{\sqrt{T_k}}\int_{T_k}{X^{k}(t)\beta_{k}(t)dt}
}{E[Y|X,Z]=\alpha+\sum_j \beta_j Z^j + \sum_k <X^k,\beta_k>}

where \eqn{Y(s)=\left[ Y^{1}(s_1),\cdots,Y^{p}(s_p)\right]}{Y(s)=[Y^1(s),...,Y^r(s)}, \eqn{Z=\left[ Z^1,\cdots,Z^p \right]}{Z=[Z^1,...,Z^p]} are the non
functional covariates, \eqn{X(t)=\left[ X^{1}(t_1),\cdots,X^{q}(t_q)
\right]}{X(t)=[X^1(t),...,X^q(t)]} are the functional ones and
\eqn{\epsilon} are random errors with mean zero , finite variance
\eqn{\sigma^2} and \eqn{E[X(t)\epsilon]=0}{E[X(t)\epsilon]=0}.  

\code{basis.y}, is a basis for represent the response. Parameter option:
\itemize{
\item Basis representation.  By default a basis of bsplines is used 
as in the \code{basis.x} parameter definition.\cr
\item Raw representation. 
The response is estimated directly using the  \code{\link{fdata}} object.
}

\code{basis.x} is a list of basis for represent each functional covariate.
The basis object can be created by the function:
\code{\link{create.pc.basis}}, \code{\link{pca.fd}}
\code{\link{create.pc.basis}}, \code{\link{create.fdata.basis}} or
\code{\link{create.basis}}.\cr \cr
}
\examples{
\dontrun{
data(tecator)
absorp <- tecator$absorp.fdata
x <- fdata.deriv(absorp)
y <- fdata.deriv(x,1)
x2 <- fdata.deriv(y)
mdat <- mfdata("y"=y,"x"=x,"x2"=x2)
# YBSP - XBSP
ff <- as.formula(y~x2)
res1 <-  fregre.lm.fr(ff,mdat)
plot(res1$beta.l$x2)

bsp7 <- create.bspline.basis(mdat$y$rangeval,7)
bsp5 <- create.bspline.basis(mdat$y$rangeval,5)
lbsp <- list("x"=bsp7,"x2"=bsp5)
res2 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=bsp7,basis.x=lbsp)
plot(res2$beta2d$x)
plot(res2$beta2d$x2)
plot(res2$beta2d$x2,type="persp",phi=30,theta=30)
# YPC - XPC
pcy <- create.pc.basis(mdat$y,1:4)
pcx <- create.pc.basis(mdat$x,1:5)
pcx2 <- create.pc.basis(mdat$x2,1:3)
lpc <- list("x"=pcx,"x2"=pcx2)
res3 <-  fregre.lm.fr(as.formula(y~x2),mdat,basis.y=bsp7,basis.x=lpc)
res3 <-  fregre.lm.fr(as.formula(y~x2),mdat,basis.y=pcy,basis.x=lpc)
res3 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=pcy,basis.x=lpc)

par(mfrow=c(1,2))
plot(res1$beta.l$x)
plot(res2$beta.l$x2)

# YPC - XBSP
pcy <- create.pc.basis(mdat$y,1:3)
res4 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=pcy)
bsp7 <- create.bspline.basis(mdat$y$rangeval,7)
bsp5 <- create.bspline.basis(mdat$y$rangeval,5)
lbsp <- list("x"=bsp7,"x2"=bsp5)
res5 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=bsp7,basis.x=lbsp)
plot(res5$beta2d$x,type="persp")
plot(res5$beta2d$x2,type="persp",phi=45,theta=60)
}

}
\references{
Chiou, J. M., Muller, H. G. and Wang, J. L. (2004). Functional response models. Statistica Sinica, 675--693.
}
\seealso{
See Also as: \code{\link{predict.fregre.lm}} and
\code{\link{summary.lm}}.\cr Alternative method: \code{\link{fregre.glm}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@usc.es}
}
\keyword{regression}
