% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.np.fr.R
\name{fregre.np.fr}
\alias{fregre.np.fr}
\title{Functional regression with functional response using non-parametric kernel
estimation}
\usage{
fregre.np.fr(
  fdataobj,
  y,
  h = NULL,
  Ker = AKer.norm,
  metric = list(x = metric.lp, y = metric.lp),
  par.metric = list(x = list(), y = list()),
  type.S = S.NW,
  par.S = list(w = 1)
)
}
\arguments{
\item{fdataobj}{\code{\link{fdata}} class object.}

\item{y}{Functional response \code{\link{fdata}} class object.}

\item{h}{Bandwidth, \code{h>0}. Default argument values are provided as the
5\%--quantile of the distance between \code{fdataobj} curves, see
\code{\link{h.default}}.}

\item{Ker}{Type of asymmetric kernel used, by default asymmetric normal
kernel.}

\item{metric}{List with components \code{x},\code{y} containing the metric 
to be applied to covariate and response, respectively.}

\item{par.metric}{List of components \code{x},\code{y} containing optional 
parameters for each metric in \code{metric}.}

\item{type.S}{Type of smothing matrix \code{S}. By default \code{S} is
calculated by Nadaraya-Watson kernel estimator (\code{S.NW}).}

\item{par.S}{List of parameters for \code{type.S}: \code{w}, the weights.}
}
\value{
Return:
\itemize{
\item {call}{ The matched call.} 
\item {fitted.values}{ Estimated scalar response.} 
\item {H}{ Hat matrix.} 
\item {residuals}{ \code{y} minus \code{fitted values}.} 
\item {df}{ The residual degrees of freedom.} 
\item {r2}{ Coefficient of determination.} 
\item {sr2}{ Residual variance.} 
\item {y}{ Response.} 
\item {fdataobj}{ Functional explanatory data.} 
\item {mdist}{ Distance matrix between \code{x} and \code{newx}.}
\item {Ker}{ Asymmetric kernel used.} 
\item {h.opt}{ Smoothing parameter or bandwidth.}
}
}
\description{
Computes functional regression between a functional explanatory variables and
functional response using kernel estimation.
}
\details{
The non-parametric functional regression model can be written as follows \deqn{y_i =r(X_i)+\epsilon_i}{ y
= r(X) + \epsilon} where the unknown smooth real function \eqn{r} is
estimated using kernel estimation by means of
\deqn{\hat{r}(X)=\frac{\sum_{i=1}^{n}{K(h^{-1}d(X,X_{i}))y_{i}}}{\sum_{i=1}^{n}{K(h^{-1}d(X,X_{i}))}}}{\hat{r}(X)=(\sum_i
K(d(X,X_i))y_i/h) / (\sum_i K(d(X,X_i)/h)) i=1,...,n} where \eqn{K} is an
kernel function (see \code{Ker} argument), \code{h} is the smoothing
parameter and \eqn{d} is a metric or a semi-metric (see \code{metric}
argument).

The distance between curves is calculated using a metric (for instance, \code{\link{metric.lp}})
or a semimetric (see \code{\link{semimetric.basis}} or \code{\link{semimetric.NPFDA}} functions).
The kernel is applied to a metric or semi-metric that provides non-negative
values, so it is common to use asymmetric kernels. Different asymmetric
kernels can be used, see \code{\link{Kernel.asymmetric}}.\cr
}
\examples{
\dontrun{
data(tecator)
ab=tecator[[1]]
ab2=fdata.deriv(ab,2)
ind=1:129
x=ab[ind]
y=ab2[ind]#' 
res.np=fregre.np.fr(x,y,Ker=AKer.epa)
summary(res.np)
}

}
\references{
Ferraty, F. and Vieu, P. (2006). \emph{Nonparametric functional
data analysis.} Springer Series in Statistics, New York. \cr

Febrero-Bande, M., Oviedo de la Fuente, M. (2012).  \emph{Statistical
Computing in Functional Data Analysis: The R Package fda.usc.} Journal of
Statistical Software, 51(4), 1-28. \url{https://www.jstatsoft.org/v51/i04/}

Hardle, W. \emph{Applied Nonparametric Regression}. Cambridge University
Press, 1994.
}
\seealso{
See Also as: \code{\link{fregre.np.cv.fr}},
\code{\link{summary.fregre.fd}} and \code{\link{predict.fregre.fd}} .\cr
Alternative method: \code{\link{fregre.basis.fr}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{regression}
