% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.gsam.fr.R
\name{fregre.sam.fr}
\alias{fregre.sam.fr}
\title{Fitting Functional Generalized Spectral Additive Models with functinal response}
\usage{
fregre.sam.fr(
  formula,
  data = list(),
  weights = NULL,
  basis.y = NULL,
  basis.x = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced
to that class): a symbolic description of the model to be fitted. The
details of model specification are given under \code{Details}.}

\item{data}{\code{list}, \code{ldata} or \code{mfdata} class object that containing the variables in the model.}

\item{weights}{weights}

\item{basis.y}{Basis for functional response.}

\item{basis.x}{List of basis for functional explanatory data estimation.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Return \code{gam} object plus:
\itemize{
\item {fitted.values}{ \code{fdata} object with the tted model predictions of expected value for each datum.} 
\item {residuals}{ \code{fdata} object with the model residuals.} 
\item {coefficients}{\code{matrix}, with coefficients of each fitted
 model (by row) and parametric coefficients (by col).}
\item {edf.coef}{\code{matrix},  with estimated degrees of freedom 
for each model fitted model (by row) and parametric coefficients (by col).
the NA value is assigned to non-significant linear and smooth terms (p-value> 0.05).} 
\item{result}{\code{list}, with each fitted \code{\link{fregre.gsam}} fitted model
for each dimension of functional response basis representation.}
\item {basis.x}{ Basis used for \code{fdata} covariates.} 
\item {basis.y}{ Basis used for \code{fdata} response.} 
\item {data}{ List that containing the variables in the model.} 
\item {formula}{ formula.} 
}
}
\description{
Computes functional GAM model between functional covariate
\eqn{(X^1(t_1),\cdots,X^{q}(t_q))}{(X(t_1),...,X(t_q))} (and non functional
covariate \eqn{(Z^1,...,Z^p)}{(Z1,...,Zp)}) and functional response
 \eqn{Y(s)=\left[ Y^{1}(s_1),\cdots,Y^{p}(s_p)\right]}{Y(s)=[Y^1(s),...,Y^r(s)}.

This function is an extension of the functional generalized additve
regression model \code{\link{fregre.gsam}} where now the response is of functional nature.

\deqn{E[Y|X,Z]) = \sum_{l=1}^{L}{f_{l}(\theta_l)}}{E[Y|X,Z]=\sum_l\sum_{l=1}{f_j(\theta_l)}} where
\eqn{\theta_l}{\xi_l}   is the coefficient of the basis function expansion of Y,  
\eqn{\xi_j^k}{\xi_j^k} is the coefficient of the basis function expansion of
\eqn{X^k}, (in PCA analysis \eqn{\xi_j^k}{\xi_j^k} is the score of the
\eqn{j}-functional PC of \eqn{X^k}) 

 \deqn{E[Y|X,Z]) = \alpha+\sum_{i=1}^{p}f_{i}(Z^{i})+\sum_{k=1}^{q}\sum_{j=1}^{k_q}{f_{j}^{k}(\xi_j^k)}}{E[Y|X,Z]=\alpha+\sum_i f_i(Z_{i})+\sum_k^q\sum_{j=1}^{k_q}{f_j^k(\xi_j^k)}}


Note that if \code{basis.y} is null, the coefficients of the basis are 
the raw values of the response observed in the grid (argvas of Y).

The smooth functions \eqn{f(\cdot)}{f(.)} can be added to the right hand
side of the formula to specify that the linear predictor depends on smooth
functions of predictors using smooth terms \code{\link{s}} and
\code{\link{te}} as in \code{\link{gam}} (or linear functionals of these as
\eqn{Z\beta} and \eqn{\big<X(t),\beta\big>}{< X(t),\beta(t) >} in
\code{\link{fregre.glm}}).

The first item in the \code{data} list is called \emph{"df"} and is a data
frame with the non functional explanatory variables, as
\code{\link{gam}}.\cr

Functional covariates and functional response of class \code{fdata} are introduced in
the following items in the \code{data} list.\cr
\code{basis.x} is a list of basis for represent each functional covariate. The basis object can be
created by the function: \code{\link{create.pc.basis}} or  \code{\link{create.fdata.basis}}.

\code{basis.y} is a  basis for represent the functional response.
 \itemize{
\item If it is \code{NULL} (by default) the response variable is not represented in any base. 
\item If it is fixed basis (\code{basisfd} class object) or data-driven basis (\code{fdata.comp}
 class for functional PCA and PLS basis)
}
}
\details{
A \code{\link{fregre.gsam}} model is adjusted for
each of the coefficients functions derived from the basis representation on a functional response.
}
\examples{
\dontrun{
data(tecator)
x=tecator$absorp.fdata
x.d1<-fdata.deriv(x)
x.d2<-fdata.deriv(x.d1)
tt<-x[["argvals"]]
dataf=as.data.frame(tecator$y)
nbasis.x=19;nbasis.y=15
basis1=create.bspline.basis(rangeval=range(tt),nbasis=nbasis.x)
basis2=create.bspline.basis(rangeval=range(tt),nbasis=nbasis.y)
basis.x=list("x"=basis1,"x.d1"=basis1)
basis.y=basis2

ldat=ldata("df"=dataf,"x"=x,"x.d1"=x.d1,"x.d2"=x.d2)
res=fregre.sam.fr(x.d2~+x+s(x.d1),ldat,
basis.x=basis.x,basis.y=basis.y)
image(res$edf.coef)
}
}
\references{
Muller HG and Stadtmuller U. (2005). \emph{Generalized
functional linear models.} Ann. Statist.33 774-805.

Wood (2001) \emph{mgcv:GAMs and Generalized Ridge Regression for R}. R News
1(2):20-25.

Ramsay, James O., and Silverman, Bernard W. (2006), \emph{ Functional Data
Analysis}, 2nd ed., Springer, New York.
}
\seealso{
See Also as: \code{\link{predict.fregre.sam.fr}}.
 Alternative methods: \code{\link{fregre.lm.fr}}
and \code{\link{fregre.kam.fr}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{regression}
