% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCV.S.fr.R
\name{GCV.S.fr}
\alias{GCV.S.fr}
\title{The generalized correlated cross-validation (GCCV) score}
\usage{
GCV.S.fr(
  y,
  S,
  criteria = "GCV",
  W = NULL,
  trim = 0,
  andf = 0,
  draw = FALSE,
  metric = metric.lp,
  par.metric = list(lp = 2)
)
}
\arguments{
\item{y}{fdata class object.}

\item{S}{Smoothing matrix, see \code{\link{S.NW}}, \code{\link{S.LLR}} or}

\item{criteria}{The penalizing function. By default \emph{"Rice"} criteria. 
Possible values are \emph{"GCCV1"}, \emph{"GCCV2"}, \emph{"GCCV3"}, \emph{"GCV"}.}

\item{W}{Matrix of weights.}

\item{trim}{The alpha of the trimming.}

\item{andf}{Degree of freedom}

\item{draw}{=TRUE, draw the curves, the sample median and trimmed mean.}

\item{metric}{Metric function, by default \code{\link{metric.lp}}.}

\item{par.metric}{Further arguments passed to metric.}
}
\value{
{ Returns GCV score calculated for input parameters.  }
}
\description{
Compute the  generalized correlated cross-validation (GCV) score.
}
\seealso{
See Also as \code{\link{fregre.np.cv.fr}} \cr Alternative method:
\code{\link{CV.S.fr}}
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{utilities}
