% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.bootstrap.R
\name{group.bootstrap}
\alias{group.bootstrap}
\title{Bootstrap for Functional Data by factor.}
\usage{
group.bootstrap(
  x,
  response,
  nb = NULL,
  smo = 0,
  Nhull = NULL,
  Nnbh = NULL,
  weights,
  metric = NULL
)
}
\arguments{
\item{x}{\code{list} that containing the variables in the model. The first item in the
\code{data} list is called \emph{"df"} and is a data frame with the response and non 
functional explanatory variables, as  \code{\link{glm}}.  Functional covariates of class}

\item{response}{\code{character}, name of the factor containing the group labels.}

\item{nb}{number of total resamples (if vector of length 1), number of resample for each class level  (if vector of length nlevels).}

\item{smo}{The smoothing parameter for the bootstrap samples as a proportion of the sample variance matrix.}

\item{Nhull}{\code{character}, size of the convex hull for generating a new sample. If \code{Nhull=0} a naive bootstrap is done.}

\item{Nnbh}{\code{character}, size of the neighborhood for selecting a local convex hull. Only needed for local convex procedures. \code{Nnbh}\eqn{>=}\code{Nhull}}

\item{weights}{Vector of weights. Only used when \code{nb} is a vector of length 1, then, the number of resamples per class is weighted.}

\item{metric}{metric to be applied to \code{x}. Only needed for local convex bootstrap, when \code{Nnbh>0}.}
}
\value{
A \code{ldata} class object.  This \code{list}  contains:
 A \code{data.frame} object (called \emph{"df"})  with the bootstrap resamples for response and multivariate variables.
 The following elements of the list are \code{fdata} class objects with bootstrap resamples for the functional variables.
}
\description{
Convex and Local Convex, smooth and naive  bootstrap for functional data (fdata and ldata class) by group label
}
\details{
{
Different types of options:
\itemize{
\item Smooth bootstrap,  \code{boot="smooth"} and \code{smo>0} (\code{Nhull} and \code{Nnbh} are not used).
\item Naive  bootstrap,  \code{boot="smooth"} and \code{smo=0} (\code{Nhull} and \code{Nnbh} are not used).
\item Global convex bootstrap,  \code{boot="convex"} and \code{Nnbh=NULL}   (\code{smo} is not used).
Select additional \eqn{X_J} elements of the same group and generate a random convex combination 
\item Local convex bootstrap,  \code{boot="convex"} and \code{Nnbh>0}   (\code{smo} is not used).  The same as previous item but now the additional elements are selected from the neighborhood of Xi in group Gi
}
}
}
\examples{
\dontrun{
data(phoneme)
mlearn <- phoneme[["learn"]]
glearn <- phoneme[["classlearn"]]
dataf <- data.frame(glearn)
dat = list("df" = dataf, "x" = mlearn)
dat.globalconvex <- group.bootstrap(dat,"glearn", boot = "convex")
dat.localconvex <- group.bootstrap(dat,"glearn", boot = "convex",Nnbh=10)
dat.smooth <- group.bootstrap(dat,"glearn", boot = "smooth", smo = 0.05)
par(mfrow=c(2,2))
plot(dat$x,col=dat$df$glearn)
plot(dat.globalconvex$x,col=dat.globalconvex$df$glearn)
plot(dat.localconvex$x,col=dat.localconvex$df$glearn)
plot(dat.smooth$x,col=dat.smooth$df$glearn)
# number of resamples by class level
dat.naive <- group.bootstrap(dat,"glearn", boot = "smooth", smo = 0, nb = c(5,5,25,5,5))
plot(dat.naive,var.name="glearn")
}
}
\seealso{
See also  \code{\link{classif.bootstrap}}
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
}
