% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldata.lagp.R
\name{ldata.lagp}
\alias{ldata.lagp}
\title{Lag a multivariate functional data (ldata or mfdata class object)}
\usage{
ldata.lagp(x, p = 2, var.name)
}
\arguments{
\item{x}{data (ldata or mfdata  class)}

\item{p}{vector of integers with he number of lags (in units of observations)}

\item{var.name}{not used yet}
}
\value{
returns a \code{ldata} with the lagged data marked with se suffixes "_p" at the end of its name.
}
\description{
Compute a lagged version of a ldata object, shifting the time base back by a given number of observations.
}
\examples{
\dontrun{ 
# no example yet
library(fda.usc.devel)
data(poblenou)
names(poblenou)
x <- ldata(poblenou$df,nox=poblenou$nox)
x0 <- ldata.lag(x,lag=0)
x5 <- ldata.lag(x,lag=5)
dim(x0[[1]])
dim(x5[[1]])
xx0 <- ldata.lagp(x,0)
xx1 <- ldata.lagp(x,1)
xx2 <- ldata.lagp(x,2)
xx12 <- ldata.lagp(x,c(1,2))
xx12 <- ldata.lagp(x,c(1,12))
names(xx0);names(xx1);names(xx2);names(xx12)
x$df[1:4,];xx0$df[1:4,];xx1$df[1:4,];xx2$df[1:4,];xx12$df[1:4,]
sapply(x,nrow);sapply(xx0,nrow);sapply(xx1,nrow);sapply(xx2,nrow);sapply(xx12,nrow)
# poner un indice que sera quen que no se repita !!!
}

}
\seealso{
\code{\link{ldata}} amd \code{\link{subset.ldata}}.
}
\author{
Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es}
}
