% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ginv.R
\name{Minverse}
\alias{Minverse}
\title{Generalized inverse}
\usage{
Minverse(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{a square numeric or complex matrix.}

\item{tol}{the tolerance for detecting linear dependencies in the columns of \code{a}. 
The default is `.Machine$double.eps.`}
}
\value{
{  Return the inverse of \code{a} }
}
\description{
This function computes the Matrix inverse.
}
\details{
By default it uses the \code{\link{solve}} function. 
If system is computationally singular, the matrix inverse is computed by
 \code{\link{svd}} function (using the effective rank).
}
\seealso{
See Also as \code{\link{solve}}, \code{\link{svd}} and \code{\link{complex}}
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{utilities}
