% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omel2008_09.R
\docType{data}
\name{omel2008_09}
\alias{omel2008_09}
\title{omel dataset}
\format{
\describe{ 
The variables are as follows:
\itemize{
\item{\code{df}:}{ \code{data.frame} with the following variables:}
\itemize{
\item{\code{index}:}{ Index of day }
\item{\code{ifecha}:}{ Date, from 2008-01-01 to 2009-12-31. }
\item{\code{dias}:}{ Day of week.}
}   
\item{\code{precio}:}{ \code{fdata} class object. The daily profiles of Electricity Price measured hourly}
\item{\code{energia}:}{ \code{fdata} class object. The daily profiles of Electricity Demand measured hourly}
}
}
}
\source{
\url{https://www.omie.es/es/market-results/daily/daily-market/daily-hourly-price}
}
\usage{
data(omel2008_09)
}
\description{
Electricity Demand and Price Data from the Iberian Energy Market
}
\details{
731  daily curves for price and volume from 2008-01-01 to 2009-12-31.
}
\references{
Febrero-Bande, M., Gonzalez-Manteiga, W., Oviedo De La Fuente, M. (2019). 
Variable selection in functional additive regression models.
Computational Statistics, 34(2), 469-487.
}
\keyword{datasets}
