% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fregre.gsam.fr.R,
%   R/predict.fregre.lm.fr.R, R/predict.fregre.mlm.fr.R
\name{predict.fregre.sam.fr}
\alias{predict.fregre.sam.fr}
\alias{predict.fregre.lm.fr}
\alias{predict.fregre.mlm.fr}
\title{Predict method for functional response model}
\usage{
\method{predict}{fregre.sam.fr}(object, newx = NULL, type = "response", ...)

\method{predict}{fregre.lm.fr}(object, newx = NULL, type = "response", ...)

\method{predict}{fregre.mlm.fr}(
  object,
  newx = NULL,
  type = "response",
  se.fit = FALSE,
  scale = NULL,
  df = NULL,
  interval = "none",
  level = 0.95,
  weights = 1,
  pred.var = res.var/weights,
  ...
)
}
\arguments{
\item{object}{\code{fregre.fr} object.}

\item{newx}{An optional data list in which to look for variables with which
to predict. If omitted, the fitted values are used. List of new explanatory
data.}

\item{type}{a character vector, Type of prediction: (\code{response}, \code{terms} for model terms or \code{effects}  for model terms where
the partial effects are summarized for each functional variable.}

\item{\dots}{Further arguments passed to or from other methods.}

\item{se.fit}{=TRUE (not default) standard error estimates are returned for
each prediction.}

\item{scale}{Scale parameter for std.err. calculation.}

\item{df}{Degrees of freedom for scale.}

\item{interval}{Type of interval calculation.}

\item{level}{Tolerance/confidence level.}

\item{weights}{variance weights for prediction. This can be a numeric vector
or a one-sided model formula. In the latter case, it is interpreted as an
expression evaluated in newdata}

\item{pred.var}{the variance(s) for future observations to be assumed for
prediction intervals. See \code{link{predict.lm}} for more details.}
}
\value{
Return the predicted values and optionally:
\itemize{
\item {predict.lm.fr}{ produces a vector of predictions
or a matrix of predictions and bounds with column names fit, lwr, and upr if
interval is set. If se.fit is TRUE, a list with the following components is
returned: fit vector or matrix as above.} 
\item {se.fit}{ standard error of predicted means.} 
\item {residual.scale}{ residual standard deviations.}
\item {df}{ degrees of freedom for residual.}
}
}
\description{
Computes predictions for regression between functional (and non functional)
explanatory variables and functional response. 
\itemize{ 
\item \code{predict.fregre.lm.fr}, Predict method for functional linear model of
\code{\link{fregre.lm.fr}} fits object using basis or principal component
representation.
\item \code{predict.fregre.mlm.fr}, Predict method for functional linear model of
\code{\link{fregre.mlm.fr}} fits object using basis or principal component
representation.
\item \code{predict.fregre.sam.fr}, Predict method for functional additive model of
\code{\link{fregre.sam.fr}} fits object using basis or principal component
representation
}

\code{predict.fregre.lm.fr} uses the model fitting function \code{\link{lm}} properties.\cr
If using functional data derived, is recommended to use a number of bases
to represent beta  lower than the number of bases used to represent the
functional data. \cr

The first item in the \code{data} list of \code{newx} argument is called
\emph{"df"} and is a data frame with the response and non functional
explanatory variables, as \code{\link{lm}}. 
Functional variables (\code{fdata} and \code{fd} class)
are introduced in the following items in the \code{data} list of \code{newx}
argument.
}
\examples{
\dontrun{
library(fda.usc.devel)
data(tecator)
absorp=tecator$absorp.fdata
ind=1:129
x=fdata.deriv(absorp[ind,])
y=fdata.deriv(x,1)
x2 <- fdata.deriv(y)
mdat<-mfdata("y"=y,"x"=x,"x2"=x2)
res0 <-  fregre.lm.fr(as.formula(y~x+x2),mdat)
pred0 <- predict.fregre.lm.fr(res0,mdat)
plot(res0$fitted.values-pred0)

res1 <-  fregre.mlm.fr(as.formula(y~x+x2),mdat)
pred1 <- predict.fregre.mlm.fr(res1,mdat)
plot(res0$fitted.values-pred0)
by<-create.bspline.basis(mdat$y$rangeval,5)
res0 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=by)
pred0 <- predict.fregre.lm.fr(res0,mdat)
plot(res0$fitted.values-pred0)

res1 <-  fregre.mlm.fr(as.formula(y~x+x2),mdat,basis.y=by)
pred1 <- predict.fregre.mlm.fr(res1,mdat)
plot(res1$fitted.values-pred0)
plot(pred1-pred0)
by<-create.pc.basis(mdat$y,1:5)
res0 <-  fregre.lm.fr(as.formula(y~x+x2),mdat,basis.y=by)
pred0 <- predict.fregre.lm.fr(res0,mdat)
plot(res0$fitted.values-pred0)

res1 <-  fregre.mlm.fr(as.formula(y~x+x2),mdat,basis.y=by)
pred1 <- predict.fregre.mlm.fr(res1,mdat)
plot(res1$fitted.values-pred0)
plot(pred1-pred0)


}                                                                                                              
}
\references{
Febrero-Bande, M., Oviedo de la Fuente, M. (2012).
\emph{Statistical Computing in Functional Data Analysis: The R Package
fda.usc.} Journal of Statistical Software, 51(4), 1-28.
\url{https://www.jstatsoft.org/v51/i04/}
}
\seealso{
See Also as: \code{\link{fregre.lm.fr}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@usc.es}
}
\keyword{regression}
