% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sp.NW.R
\name{Sp.NW}
\alias{Sp.NW}
\title{Smoothing matrix for multivariate data}
\usage{
Sp.NW(x, h = 0.1, w = NULL, cv = FALSE)
}
\arguments{
\item{x}{\code{matrix}, Multivariate dataset}

\item{h}{Smoothing parameter or bandwidth. If \code{numeric} a fixed bandidth applied to the variance and covariance matrix of X.  If \code{matrix} is the bandwidth matrix.}

\item{w}{Optional case weights.}

\item{cv}{If \code{TRUE}, cross-validation is done (see \code{\link{CV.S}} and \code{\link{GCV.S}}).}
}
\value{
Return the smoothing matrix \code{S} by Nadaraya-Watson estimator using Gaussian kernel.
}
\description{
Provides the smoothing matrix \code{S} for multivariate data using a Gaussian Kernel
}
\details{
Options: 
\itemize{
 \item Nadaraya-Watson kernel estimator (S.NW) with bandwidth parameter \code{h}. 
 }
}
\examples{
\dontrun{
data(tecator)
x <- tecator$absorp.fdata$data[,c(1,25,50,75,100)]
H1 <- Sp.NW(x, cv = T)
H2 <- Sp.NW(x, cv = F)
y <- tecator$y$Fat
yest1 <- H1 \%*\% y
yest2 <- H2 \%*\% y
plot(y,yest1,pch=19)
points(y,yest2,col=2,pch=19)
}
}
\references{
Scott, D. W. (2015). \emph{Multivariate density estimation: theory, practice, and visualization.} John Wiley \& Sons.
}
\seealso{
See Also \code{\link{S.np}} for functional data object and  \code{\link{fregre.np}} for  non-parametric regression wit scalar response.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es}
}
\keyword{smooth}
