% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fregre.fr.r
\name{summary.fregre.fr}
\alias{summary.fregre.fr}
\title{Summarizes information from fregre.fr objects.}
\usage{
\method{summary}{fregre.fr}(object, times.influ = 3, draw = TRUE, ...)
}
\arguments{
\item{object}{Estimated by functional response regression, \code{fregre.fr} object.}

\item{times.influ}{Limit for detect possible infuence curves.}

\item{draw}{=TRUE draw estimation and residuals graphics.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
Summary function for  \code{\link{fregre.basis.fr}},
\code{\link{fregre.np.fr}} and \code{\link{fregre.np.cv.fr}} functions.

Shows:\cr \tabular{ll}{ \tab -Call.\cr \tab -R squared.\cr \tab -Residual
variance.\cr \tab -Index of possible atypical curves or possible}

If draw=TRUE plot: \cr \tabular{ll}{ \tab -y vs y fitted values.\cr \tab
-Residuals vs fitted values.\cr  -Functional residual boxplot.\cr }
 If \code{ask}=FALSE draw graphs in one window, by default. If \code{ask}=TRUE, draw each graph in a window,
waiting to confirm.
}
\seealso{
\code{\link{fregre.basis.fr}}, \code{\link{fregre.np.fr}} and \code{\link{fregre.np.cv.fr}}.
}
\author{
Manuel Febrero-Bande and Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es}
}
\keyword{print}
